requirements.txt
cmake==4.1.0
precomp_whl\dlib-19.24.99-cp312-cp312-win_amd64.whl

click==8.3.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==11.3.0
polars==1.33.1
protobuf==4.25.8
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
supervision==0.26.1
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
urllib3==2.5.0
pillow==11.3.0